# SmartParking_API_Testing/OpenAPI Specification/api-docs.json
{"openapi":"3.0.1","info":{"title":"Smart Parking Building API","description":"Hệ thống quản lý nhà xe thông minh - Swagger UI API Documentation","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/vehicles/{id}":{"get":{"tags":["vehicle-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"put":{"tags":["vehicle-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"delete":{"tags":["vehicle-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/driver/vehicle/lock":{"put":{"tags":["driver-controller"],"operationId":"lockVehicle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LockVehicleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/sessions/{id}":{"get":{"tags":["parking-session-controller"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ParkingSession"}}}}}},"put":{"tags":["parking-session-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingSession"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ParkingSession"}}}}}},"delete":{"tags":["parking-session-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/vehicles":{"get":{"tags":["vehicle-controller"],"operationId":"all","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"post":{"tags":["vehicle-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/api/v1/parking/verify-exit-qr":{"post":{"tags":["parking-controller"],"operationId":"verifyExitQr","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyExitQrRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/parking/congestion/checkout":{"post":{"tags":["parking-controller"],"operationId":"congestionCheckout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CongestionCheckoutRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}}},"/api/v1/parking/checkout/{cardId}":{"post":{"tags":["parking-controller"],"operationId":"checkout","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}}},"/api/v1/parking/check-in/visitor":{"post":{"tags":["parking-controller"],"operationId":"visitorCheckIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitorCheckInRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckInResponse"}}}}}}},"/api/v1/parking/check-in/ai":{"post":{"tags":["parking-controller"],"operationId":"aiCheckIn","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiCheckInRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckInResponse"}}}}}}},"/api/v1/driver/qr/generate":{"post":{"tags":["driver-controller"],"operationId":"generateQr","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateQrRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VipQrIdentifier"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"all_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/sessions":{"get":{"tags":["parking-session-controller"],"operationId":"all_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ParkingSession"}}}}}}},"post":{"tags":["parking-session-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingSession"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ParkingSession"}}}}}}},"/api/blacklisted-cards":{"get":{"tags":["blacklist-card-controller"],"operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlacklistEntry"}}}}}}},"post":{"tags":["blacklist-card-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlacklistCardRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlacklistEntry"}}}}}}},"/api/blacklisted-cards/block-by-plate":{"post":{"tags":["blacklist-card-controller"],"operationId":"blockCardByLicensePlate","parameters":[{"name":"plate","in":"query","required":true,"schema":{"type":"string"}},{"name":"reason","in":"query","required":true,"schema":{"type":"string"}},{"name":"staffId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlacklistEntry"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/blacklisted-cards/check/{cardId}":{"get":{"tags":["blacklist-card-controller"],"operationId":"check","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/blacklisted-cards/{id}":{"delete":{"tags":["blacklist-card-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Vehicle":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"ownerId":{"type":"string","format":"uuid"},"licensePlate":{"type":"string"},"vehicleSize":{"type":"string"},"color":{"type":"string"},"colorRgb":{"type":"string"},"bodyShape":{"type":"string"},"brand":{"type":"string"},"registrationDocUrl":{"type":"string"},"registrationPhotoUrl":{"type":"string"},"violationCount":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"active":{"type":"boolean"}}},"LockVehicleRequest":{"type":"object","properties":{"vehicleId":{"type":"string","format":"uuid"},"lockStatus":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"passwordHash":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string","enum":["ADMIN","MANAGER","STAFF","DRIVER"]},"status":{"type":"string","enum":["ACTIVE","INACTIVE","SUSPENDED"]},"fcmToken":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ParkingSession":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"licensePlate":{"type":"string"},"vehicleId":{"type":"string","format":"uuid"},"cardId":{"type":"string","format":"uuid"},"validatedQrId":{"type":"string","format":"uuid"},"assignedZoneId":{"type":"string","format":"uuid"},"parkedSlotId":{"type":"string","format":"uuid"},"checkInTime":{"type":"string","format":"date-time"},"checkOutTime":{"type":"string","format":"date-time"},"sessionStatus":{"type":"string","enum":["ACTIVE","COMPLETED","PASSED_CONFIRMED","LOST_CARD"]},"isVip":{"type":"boolean"},"isLocked":{"type":"boolean"},"isSuspicious":{"type":"boolean"},"suspiciousReason":{"type":"string"},"overrideByStaff":{"type":"string","format":"uuid"},"overrideReason":{"type":"string"},"mobileCheckoutStaffId":{"type":"string","format":"uuid"},"mobileCheckoutLocation":{"type":"string"},"mobileCheckoutAt":{"type":"string","format":"date-time"},"mobileCheckoutPhoto":{"type":"string"},"lostCardProofPhotos":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"VerifyExitQrRequest":{"type":"object","properties":{"detectedPlate":{"type":"string"},"qrToken":{"type":"string"}}},"CongestionCheckoutRequest":{"type":"object","properties":{"licensePlate":{"type":"string"},"staffId":{"type":"string","format":"uuid"},"gpsLocation":{"type":"string"},"proofImageUrl":{"type":"string"},"paymentMethod":{"type":"string","enum":["CASH","VNPAY_SANDBOX","MOMO_SANDBOX","QR_BANK"]}}},"Transaction":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sessionId":{"type":"string","format":"uuid"},"parkingFee":{"type":"number"},"lostCardPenalty":{"type":"number"},"violationPenalty":{"type":"number"},"totalAmount":{"type":"number"},"paymentMethod":{"type":"string","enum":["CASH","VNPAY_SANDBOX","MOMO_SANDBOX","QR_BANK"]},"paymentStatus":{"type":"string","enum":["PENDING","SUCCESS","FAILED","REFUNDED"]},"paymentReference":{"type":"string"},"processedBy":{"type":"string","format":"uuid"},"isMobileCheckout":{"type":"boolean"},"mobileGpsLocation":{"type":"string"},"mobilePhotoProof":{"type":"string"},"receiptUrl":{"type":"string"},"processedAt":{"type":"string","format":"date-time"}}},"VisitorCheckInRequest":{"type":"object","properties":{"plate":{"type":"string"},"vehicle_type":{"type":"string"},"card_code":{"type":"string"},"image_url":{"type":"string"}}},"CheckInResponse":{"type":"object","properties":{"session_id":{"type":"string"},"assigned_zone_code":{"type":"string"},"message":{"type":"string"}}},"AiCheckInRequest":{"type":"object","properties":{"plate":{"type":"string"},"vehicle_type":{"type":"string"},"image_url":{"type":"string"},"camera_id":{"type":"string"},"confidence_score":{"type":"number","format":"double"}}},"GenerateQrRequest":{"type":"object","properties":{"vehicleId":{"type":"string","format":"uuid"},"purpose":{"type":"string"}}},"VipQrIdentifier":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"vehicleId":{"type":"string","format":"uuid"},"qrToken":{"type":"string"},"purpose":{"type":"string"},"expiredAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"used":{"type":"boolean"}}},"BlacklistCardRequest":{"type":"object","properties":{"cardId":{"type":"string","format":"uuid"},"sessionId":{"type":"string","format":"uuid"},"reason":{"type":"string"},"blacklistedBy":{"type":"string","format":"uuid"},"notes":{"type":"string"}}},"BlacklistEntry":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"cardId":{"type":"string","format":"uuid"},"sessionId":{"type":"string","format":"uuid"},"reason":{"type":"string"},"blacklistedBy":{"type":"string","format":"uuid"},"blacklistedAt":{"type":"string","format":"date-time"},"notes":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}